created one file to read from database
